BUT ACTUALLY DOE, FACEBOOK EVENT INTEGRATION
Creating an event on our app creates a corresponding event on Facebook. It's pretty badass. The code for it is a fucking mess of callbacks and shit but I'm working on that so don't worry. Also it'll all be commented soon.

TODO: pics for events (default to whatever pic corresponds to google place?)